requirements.txt
# vllm==0.10.2  ## manually install vllm from local code source
descript-audiotools==0.7.2
matplotlib==3.8.2
omegaconf
sentencepiece
librosa
soundfile
gradio
ninja
modelscope
munch==4.0.0
loguru

# should manually install openfst-1.8.3 (for pynini-2.1.6) before installing pynini

WeTextProcessing; platform_machine != "Darwin"
wetext; platform_system == "Darwin"

numpy<2     ## add for compatibility with other packages
torchcodec  ## add for error "ModuleNotFoundError: No module named 'torchcodec'"
